requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
boto3==1.36.22
botocore==1.36.22
cachetools==6.2.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dateparser==1.2.2
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
fastapi==0.135.1
filelock==3.20.0
Flask==3.1.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Levenshtein==0.27.3
lightgbm==4.6.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
narwhals==2.15.0
networkx==3.5
numba==0.64.0
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.1
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.3
pyarrow==22.0.0
PyAutoGUI==0.9.54
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.20.0
PyMsgBox==1.0.9
pyparsing==3.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==9.0.3
python-binance==1.0.34
python-can==4.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-Levenshtein==0.27.3
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.4
rpds-py==0.30.0
s3transfer==0.11.2
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
shap==0.51.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
starlette==0.52.1
streamlit==1.55.0
streamlit-aggrid==1.2.1.post2
sympy==1.14.0
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
tk==0.1.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.42.0
vaderSentiment==3.3.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xgboost==3.2.0
XlsxWriter==3.2.2
yarl==1.22.0
